% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTestRev.R
\name{cor.test_trimmed_v0.default}
\alias{cor.test_trimmed_v0.default}
\title{cor.test_trimmed_v0.default
Trimmed cor test that is faster than standard cor.test}
\usage{
\method{cor.test_trimmed_v0}{default}(
  x,
  y,
  alternative = c("two.sided", "less", "greater"),
  method = c("pearson"),
  exact = NULL,
  continuity = FALSE,
  ...
)
}
\description{
cor.test_trimmed_v0.default
Trimmed cor test that is faster than standard cor.test
}
