% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTestRev.R
\name{cor.test_trimmed_v0}
\alias{cor.test_trimmed_v0}
\title{cor.test_trimmed_v0
Trimmed cor test that is faster than standard cor.test}
\usage{
cor.test_trimmed_v0(x, ...)
}
\description{
cor.test_trimmed_v0
Trimmed cor test that is faster than standard cor.test
}
