% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationBetCrisprDrugRev.R
\name{CorCRISPRDrugR}
\alias{CorCRISPRDrugR}
\title{Compute a correlation between the every gene crispr KO vs each drug response}
\usage{
CorCRISPRDrugR(
  infunc_drugName = drugname,
  infunc_drugResponse = drugResponse,
  infunc_CRISPRResponse = CRISPRResponse
)
}
\arguments{
\item{infunc_drugName}{Drug Name}

\item{infunc_drugResponse}{scores from the drug}

\item{infunc_CRISPRResponse}{scores from the KO metnohd}
}
\value{
correlation_df  Genes with corelation values and P val associated with drug across the same celllines
}
\description{
to compute the correlations between the viability of cell lines after CRISPR knockout of each gene and the viability of the same cell lines after drug treatment
}
\details{
CorCRISPRDrugR
}
\examples{
data (OntargetM)
all_drugs <- OntargetM$DrugMetadata[,"broad_id_trimmed"]
sample_drugs=sample(all_drugs, 5)
metadata.f <- OntargetM$DrugMetadata[sample_drugs,]
drug.prism.f <- OntargetM$secondary_prism[sample_drugs,]
drugVScrispr.Corr.Features.List=mclapply(sample_drugs,function(x) CorCRISPRDrugR(x, infunc_drugResponse=drug.prism.f, infunc_CRISPRResponse=OntargetM$avana_CRISPR),mc.cores = 2)
names (drugVScrispr.Corr.Features.List) <- sample_drugs
saveRDS(drugVScrispr.Corr.Features.List,"drugVScrispr.Corr.Features.List.RDS")
}
