% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKSKnownTarget.R
\name{DKSKnownTarget}
\alias{DKSKnownTarget}
\title{provide a similarity score ( corelation values) for Known Target of a drug and the best Primary Target for each one.}
\usage{
DKSKnownTarget(
  infunc_drugVScrispr_corr_features_list = drugVScrispr_corr_features_list,
  infunc_drug_metadata = metadata.Drug
)
}
\arguments{
\item{infunc_drugVScrispr_corr_features_list}{The list of result from "CorCRISPRDrugR" function.}

\item{infunc_drug_metadata}{Meta data from the drug.}
}
\value{
knownTarget_predictions object: The list of targeted genes along with drug information and  "best corelation" and "max corelation values" along with P values
}
\description{
It takes the name of a drug as an input and returns a list containing the DKS score and P-value for the correlation between the drug and each gene.
}
\details{
DKSKnownTarget
}
\examples{
data (OntargetM)
library (parallel)
all_drugs <- OntargetM$DrugMetadata[,"broad_id_trimmed"]
sample_drugs=sample(all_drugs, 5)
metadata.f <- OntargetM$DrugMetadata[sample_drugs,]
drug.prism.f <- OntargetM$secondary_prism[sample_drugs,]
drugVScrispr.Corr.Features.List=mclapply(sample_drugs,function(x) CorCRISPRDrugR (x, infunc_drugResponse=drug.prism.f, infunc_CRISPRResponse=OntargetM$avana_CRISPR),mc.cores = 2)
names (drugVScrispr.Corr.Features.List) <- sample_drugs
KnownTarget.predictions <- DKSKnownTarget( infunc_drugVScrispr_corr_features_list=drugVScrispr.Corr.Features.List,infunc_drug_metadata = metadata.f)
saveRDS(KnownTarget.predictions, "KnownTarget_predictions.RDS")
}
