% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSpecificityScore.R
\name{ExprSpecificityScore}
\alias{ExprSpecificityScore}
\title{Compute interaction between the drug and KO expression in term of lower vs higher expression}
\usage{
ExprSpecificityScore(
  infunc_KnownTarget = KnownTarget_predictions,
  infunc_Expression = expression_matched,
  infunc_drugResponse = drug.Res.PRISM.matched,
  infunc_CRISPRResponse = CRISPR.KO.Res.matched,
  infunc_low_ex_cut_off = 3
)
}
\arguments{
\item{infunc_KnownTarget}{The list of targeted genes along with drug informaiton and corelation values "best corelation" and max corelation values along with P values}

\item{infunc_Expression}{Expression matrix}

\item{infunc_drugResponse}{Drug scores.}

\item{infunc_CRISPRResponse}{scores from KO method}

\item{infunc_low_ex_cut_off}{desired cut-off for low expression}
}
\description{
provide each gene a likelihood score of being a secondary target of drug
}
\examples{
data (OntargetM)
KnownTarget.predictions <- readRDS ( "KnownTarget_predictions.RDS")
Low_expression_cellLines = sapply(KnownTarget.predictions$MaxTargetName, function(x) err_handle(names(which(dat.Expr.matched[x,]<2))))
interaction.Features.Secondary.LowExpr <- ExprSpecificityScore (infunc_KnownTarget=KnownTarget.predictions,infunc_Expression=OntargetM$expression_20Q4,infunc_drugResponse=OntargetM$secondary_prism, infunc_CRISPRResponse=OntargetM$avana_CRISPR,infunc_low_ex_cut_off  = 2)
TargetLoweEpr.specificity=data.frame(LowExpr_interaction_strength=sapply(interaction.Features.Secondary.LowExpr, function(x) x[1]),
LowExpr_interaction_P=sapply(interaction.Features.Secondary.LowExpr, function(x) x[2]))
identical ( names(interaction.Features.Secondary.LowExpr), KnownTarget.predictions$drugName)
TargetLoweEpr.specificity=cbind(KnownTarget.predictions[,1:2], TargetLoweEpr.specificity)
saveRDS(TargetLoweEpr.specificity, file='Target_LowedEpr_specificity.RDS')
}
