% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutantSpecificityScore.R
\name{MutantSpecificityScore}
\alias{MutantSpecificityScore}
\title{Compute interaction between the drug and KO expression in term of mutant}
\usage{
MutantSpecificityScore(
  infunc_KnownTarget = KnownTarget_predictions,
  infunc_mutant = dat.Mutation.mached,
  infunc_drugResponse = drug.Res.PRISM.matched,
  infunc_CRISPRResponse = CRISPR.KO.Res.matched
)
}
\arguments{
\item{infunc_KnownTarget}{The list of targeted genes along with drug information and corelation values "best corelation" and max corelation values along with P values}

\item{infunc_mutant}{Mutation matrix ( zero and 1)}

\item{infunc_drugResponse}{Drug scores.}

\item{infunc_CRISPRResponse}{scores from KO method}
}
\description{
compute the Mutant Specificity Score (MS Score) of a drug to its known target
}
\details{
MutantSpecificityScore
}
\examples{
data (OntargetM)
KnownTarget.predictions <- readRDS ( "KnownTarget_predictions.RDS")
interaction.Features.Mutant<- MutantSpecificityScore (infunc_KnownTarget=KnownTarget.predictions,infunc_mutant=OntargetM$mutations_mat,infunc_drugResponse=OntargetM$secondary_prism, infunc_CRISPRResponse=OntargetM$avana_CRISPR)
Target_Mutation_specificity=data.frame(mutation_interaction_strength=sapply(interaction.Features.Mutant, function(x) x[1]), mutation_interaction_P=sapply(interaction.Features.Mutant, function(x) x[2]))
Target_Mutation_specificity=cbind(KnownTarget_predictions[,1:2], Target_Mutation_specificity)
saveRDS(Target_Mutation_specificity,'Target_Mutation_specificity.RDS');
}
