% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OntargetM}
\alias{OntargetM}
\title{OntargetM. #' an object containing Viability matrix after CRISPR-KO; Viability after Drug Treatment; Drug metadata from Broad, mutation matrix, and expression matrix with common celllines and common drugs.
The list are as follows:}
\format{
the format is the list of 5 \code {matrix} \itemize{
\item\$DrugMetadata:\code{matrix}  containing 1448 drugs with 5 columns:  broad_id_trimmed, name, target, drug_category, and moa
\item\$secondary_prism: \code{matrix} containing Viability scores after Drug Treatment with 1448 drugs across 319 celllines
\item\$avana_CRISPR: \code{matrix} containing Viability scores after CRISPR-KO with 18333 genes across 319 celllines
\item\$mutations_mat:\code{matrix} containing Viability scores after CRISPR-KO with 19350 genes across 319 celllines
\item\$expression_20Q4:\code{matrix} containing Viability scores after CRISPR-KO with 19182 genes across 319 celllines. }
DrugMetadata

This dataset was from Corsello_supplemental_tables.xlsx:
}
\source{
depmap \url{ https://depmap.org/repurposing/#:~:text=Corsello_supplemental_tables.xlsx}

depmap \url{https://depmap.org/portal/download/all/?releasename=PRISM+Repurposing+19Q4&filename=secondary-screen-dose-response-curve-parameters.csv}

depmap \url{https://depmap.org/portal/download/all/?releasename=DepMap+Public+22Q4&filename=CRISPRGeneEffect.csv}
mutations_mat
Mutation binary profile for ~2K cell lines; 0 is WT; 1 is mutated

depmap \url{https://depmap.org/portal/download/all/?releasename=DepMap+Public+22Q4&filename=OmicsSomaticMutations.csv}

expression_20Q4
Expression profile of ~2K cell lines for 20K genes; file name: OmicsExpressionProteinCodingGenesTPMLogp1.csv

depmap \url{https://depmap.org/portal/download/all/}
}
\description{
OntargetM. #' an object containing Viability matrix after CRISPR-KO; Viability after Drug Treatment; Drug metadata from Broad, mutation matrix, and expression matrix with common celllines and common drugs.
The list are as follows:
}
\keyword{(}
\keyword{1,448}
\keyword{20K}
\keyword{489}
\keyword{8-step:}
\keyword{CRISPR}
\keyword{Corsello_supplemental_tables.xlsx}
\keyword{Drug}
\keyword{KO}
\keyword{Secondary_prism}
\keyword{Treatment}
\keyword{against}
\keyword{an}
\keyword{avana_CRISPR}
\keyword{cell}
\keyword{drugs}
\keyword{for}
\keyword{genes}
\keyword{in}
\keyword{lines}
\keyword{of}
\keyword{profile}
\keyword{scores)}
\keyword{secondary-screen-dose-response-curve-parameters.csv}
\keyword{~}
\keyword{~800}
