% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimaryTargetPathwayResolution.R
\name{Primary_Target_at_Pathway_Resolution}
\alias{Primary_Target_at_Pathway_Resolution}
\title{Extract primary Target at a pathway Level; Provides a score for each pathway to be the MOA of a drug"}
\usage{
Primary_Target_at_Pathway_Resolution(
  infunc_drugVScrispr_corr_features_list = drugVScrispr_corr_features_list,
  infunc_drug_metadata = metadata.Drug
)
}
\arguments{
\item{infunc_drugVScrispr_corr_features_list}{The list of result from "correlation_bet_crispr_drug_r" function.}

\item{infunc_drug_metadata}{meta data from drug}
}
\value{
Pathway level probability to be a primary MOA; gsea_enrichment_all_drugs
}
\description{
Predicts a Primary Target at a pathwya Level. It next finds the pathways that are most enriched in the genes with high DKS scores. It does this by performing a pathway enrichment test on the ranked gene list by DKS score. The output is a matrix of pathway-level probabilities for each drug to be the primary MOA
}
\examples{
data (OntargetM)
all_drugs <- OntargetM$DrugMetadata[,"broad_id_trimmed"]
sample_drugs=sample(all_drugs, 5)
metadata.f <- OntargetM$DrugMetadata[sample_drugs,]
drug.prism.f <- OntargetM$secondary_prism[sample_drugs,]
drugVScrispr.Corr.Features.List=mclapply(sample_drugs,function(x) correlation_bet_crispr_drug_r(x, infunc_drugResponse=drug.prism.f, infunc_CRISPRResponse=OntargetM$avana_CRISPR),mc.cores = 2)
names (drugVScrispr.Corr.Features.List) <- sample_drugs
gsea_enrichment <- Primary_Target_at_Pathway_Resolution ( infunc_drugVScrispr_corr_features_list=drugVScrispr.Corr.Features.List,infunc_drug_metadata = metadata.f)
}
